Move TagsApiConsumer to web requester folder
