+++ Editor/reup.romulo.Editor.asmdef
@@ -9,7 +9,6 @@
         "GUID:09fefbc308747e34c917406ce5dd7025",
         "GUID:51949aeab7ddd114ea276b96726cb4bb",
         "GUID:5f723ba0ce93eb64ca6f59169571c5ed",
-        "GUID:b3e03b714bd206c43be1c14e98aa81df",
         "GUID:3e29f3cf4a29b6f4a8ecdaccb3a3a2ba",
         "GUID:1772869efacc54544ac5d329d23fc591",
         "GUID:0da6763f50b07ec4c9363bdf3985cee5",
@@ -18,7 +17,9 @@
         "GUID:993f0e698e8fb354b8eee1296bc5f2f6",
         "GUID:f33eeee41b733464e94ee9a139b85f8a",
         "GUID:db1051acd1be67444bdd582f8c8da4bd",
-        "GUID:ef71f4f9d4e9ed44987e6a7778893414"
+        "GUID:ef71f4f9d4e9ed44987e6a7778893414",
+        "GUID:2fff08516fe8b484ea0191f9747e8d2b",
+        "GUID:f8a37e23801ddf543939c4801a49d87e"
     ],
     "includePlatforms": [
         "Editor"
