+++ Runtime/ManagerInterfaces/ITagsApiManager.cs
@@ -1,4 +1,4 @@
-using ReupVirtualTwin.controllerInterfaces;
+using ReupVirtualTwin.webRequestersInterfaces;
 using ReupVirtualTwin.dataModels;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -7,7 +7,7 @@ namespace ReupVirtualTwin.managerInterfaces
 {
     public interface ITagsApiManager
     {
-        public ITagsWebRequesterController webRequester { get; set; }
+        public ITagsApiConsumer webRequester { get; set; }
         public Task<List<ObjectTag>> GetTags();
         public Task<List<ObjectTag>> LoadMoreTags();
         public void CleanTags();
