+++ Runtime/ManagerInterfaces/reup.romulo.managerInterfaces.asmdef
@@ -5,7 +5,8 @@
         "GUID:3e29f3cf4a29b6f4a8ecdaccb3a3a2ba",
         "GUID:ef71f4f9d4e9ed44987e6a7778893414",
         "GUID:a94387ec035d5514c93dfbc1c4e15171",
-        "GUID:15c1dcf2b670848439ceff09bf17acd2"
+        "GUID:15c1dcf2b670848439ceff09bf17acd2",
+        "GUID:2fff08516fe8b484ea0191f9747e8d2b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
