+++ Runtime/Managers/reup.modulo.managers.asmdef
@@ -13,7 +13,8 @@
         "GUID:a94387ec035d5514c93dfbc1c4e15171",
         "GUID:8d6c531fc763e1c46bbfd59fec526b43",
         "GUID:5f723ba0ce93eb64ca6f59169571c5ed",
-        "GUID:15c1dcf2b670848439ceff09bf17acd2"
+        "GUID:15c1dcf2b670848439ceff09bf17acd2",
+        "GUID:2fff08516fe8b484ea0191f9747e8d2b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
