+++ Runtime/WebRequestersInterfaces/ITagsApiConsumer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 502919847dcd12c4ab4af2679fd54d95
+guid: 6e283207ec0fedb4f94eb4376e45e495
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
